+++ tests/test_btc_monitor.py
@@ -18,20 +18,42 @@ async def test_async_monitoring(btc_client: AioTxBTCClient):
     async def handle_transaction(transaction):
         transactions.append(transaction)
 
-    await btc_client.import_address("tb1pdgattqpwrzvrjhwkv0hs5lg6nkwwjpp48mhuh9mc259ftqagryyqvlv4we", 555)
+    await btc_client.import_address("tb1pmuuv2qjujv9qawqcc424nhlnmux7mvsyqj7qgc6z0vwqujvx4k9s34kuxn", 2811502)
+    await btc_client.import_address("tb1putn7zkfjr97xd77as09w5syy3x7xr5crp99wycjcxgdwrq3le9lse9fdkd")
+    await btc_client.import_address("tb1p2z70my9zwxsuqevdxjn909c7jyp89z4qv3uhy5wggn47ruvfa65s4fnngd")
+    await btc_client.import_address("tb1paf6damf5052arl3r2lsufuhyu48yth8mrgxdtqj9pervz624q6xqxm7ew3")
+    await btc_client.monitor._add_new_utxo("tb1paf6damf5052arl3r2lsufuhyu48yth8mrgxdtqj9pervz624q6xqxm7ew3", "8b88ebe1d0bb1eda7b1512853e75d3884b99e6634317589b1f7457ab92f1e41f", 499824213,0)
+    start_balance_utxo_wallet = await btc_client.get_balance("tb1paf6damf5052arl3r2lsufuhyu48yth8mrgxdtqj9pervz624q6xqxm7ew3")
+    assert start_balance_utxo_wallet == 499824213
     await btc_client.start_monitoring()
     
+    
     try:
-        await asyncio.sleep(3)
+        await asyncio.sleep(2.5)
     except KeyboardInterrupt:
         btc_client.stop_monitoring()
 
     assert len(blocks) > 0
     assert len(transactions) > 0
 
-    assert 555 in blocks
-    assert 556 in blocks
-    assert "070a808197d97c8982587f75fef6d531282863a110ad5e8b42cdec8ddb6dc4e0" in [tx["txid"] for tx in transactions]
-    assert "c852040a873a8f3c1ee69fda40dcf32a118978a923f62176272e840fc042de54" in [tx["txid"] for tx in transactions]
+    balance = await btc_client.get_balance("tb1pmuuv2qjujv9qawqcc424nhlnmux7mvsyqj7qgc6z0vwqujvx4k9s34kuxn")
+    assert balance == 333
+
+    balance = await btc_client.get_balance("tb1putn7zkfjr97xd77as09w5syy3x7xr5crp99wycjcxgdwrq3le9lse9fdkd")
+    assert balance == 36731406
+
+    balance = await btc_client.get_balance("tb1p2z70my9zwxsuqevdxjn909c7jyp89z4qv3uhy5wggn47ruvfa65s4fnngd")
+    assert balance == 478837110
+
+
+    end_balance_utxo_wallet = await btc_client.get_balance("tb1paf6damf5052arl3r2lsufuhyu48yth8mrgxdtqj9pervz624q6xqxm7ew3")
+    assert end_balance_utxo_wallet == 0
+
+    assert 2811502 in blocks
+    assert 2811503 in blocks
+    assert "e6eda82369b59c805998fc554f55163d4e4d38f5817adf8c1b6b9f917937bd77" in [tx["txid"] for tx in transactions]
+    assert "56f677b37c44f9609f6d2ee121effe5a59b0c550c9be5073d03d8c365dde0e83" in [tx["txid"] for tx in transactions]
+
+